Buyprevious say command bug fixed!
The bug? if someone bought an item or had a few items already before they used the buyprevious command, they would be able to go beyond the max allowed items allowed for any server running this bug.

The fix was to remove only the items a player did not own when they died, then buy only the items they had when they died.  This will also help not to overcharge a player.  It also keeps tabs not to buy items they can not afford to buy.  It lets the player know which items it discards, and buys.

git-svn-id: http://ownageclan.com/svn/war3source@798 e104397d-def3-435e-8314-01917568762b